fix: moare project patching

+++ ci/patch-ios-project.js
@@ -7,6 +7,9 @@ post_install do |installer|
   installer.pods_project.targets.each do |target|
     target.build_configurations.each do |config|
       config.build_settings.delete 'IPHONEOS_DEPLOYMENT_TARGET'
+      config.build_settings['EXPANDED_CODE_SIGN_IDENTITY'] = ""
+      config.build_settings['CODE_SIGNING_REQUIRED'] = "NO"
+      config.build_settings['CODE_SIGNING_ALLOWED'] = "NO"
     end
   end
 end
@@ -58,9 +61,14 @@ module.exports = context => {
 
     const podFilePath = path.join(projectDir, 'Podfile')
 
-    const podFileContent = fs.readFileSync(path.join(projectDir, 'Podfile'));
+    if ( fs.existsSync(projectDir) ) {
+      console.log('Patching Podfile!');
+      const podFileContent = fs.readFileSync(path.join(projectDir, 'Podfile'));
 
-    fs.writeFile(podFilePath, podFileContent + podFilePatch)
+      fs.writeFile(podFilePath, podFileContent + podFilePatch)
+    } else {
+      console.error('Could not find Podfile');
+    }
 
     const project = xcode.project(projectPath);
